Added --cookie feature
You can now supply your Slack 'd' cookie to the script using the --cookie flag. This will look for all workspaces you're a member of and print the URL's and associated tokens. The tokens can then be used in the next part of the attack by using the --token flag.