Minor changes
I renamed some of the variables so that they are easier (for me anyway) to understand.

Also removed the `if (anyTrue and anyFalse)` as that condition will actually never be hit as specifying --no-flag and --flag <--- the latter flag will take precedence.

It's easier to just remove the check completely but if you guys think it's worth doing a check for contradictory flags then we can do that too but it could get complicated.

Otherwise the default behaviour will mean that the 'latest' flag gets set and I don't mind that.

Thoughts @milangfx @influenza